fix(arch): resolve all test_no_private_exports architecture violations
- Gateway: remove _Authn/_Fwd aliases from bootstrap/__init__.py,
  move _inject_enterprise_plugins to _container.py
- BaaS: remove private function re-exports from __all__ in
  config_manage, device_manage, template_manage, tenant_manage
- BaaS: move _is_expected_distributed_lock_conflict to _utils.py
- BaaS: extract BareLoggerPlugin to _plugin.py (logger/bare)
- BaaS: rename _BotCreateResult → BotCreateResult et al in teclaw SPI
- BaaS: move _inject_enterprise_plugins to _container.py
- Clear _KNOWN_PRIVATE_EXPORTS allowlists in both test files
- Update test imports for renamed/moved symbols

Closes #779
