+++ src/baas/src/secbaas/community/core/service/config_manage/__init__.py
@@ -2,10 +2,8 @@
 
 from ._system_config_service import (
     DefaultSystemConfigManageService,
-    _record_to_response,
 )
 
 __all__ = [
     "DefaultSystemConfigManageService",
-    "_record_to_response",
 ]
