+++ src/baas/src/secbaas/community/core/service/device_manage/__init__.py
@@ -2,9 +2,6 @@
 
 from ._device_service import (
     DefaultDeviceService,
-    _decrypt_header_rule_values,
-    _encrypt_header_rule_values,
-    _safe_format_hook,
     common_sm4_decrypt,
     common_sm4_encrypt,
     device_record_to_response,
@@ -18,8 +15,5 @@ __all__ = [
     "common_sm4_encrypt",
     "dispatch_start_hook",
     "get_current_env",
-    "_decrypt_header_rule_values",
-    "_encrypt_header_rule_values",
     "device_record_to_response",
-    "_safe_format_hook",
 ]
