+++ src/baas/src/secbaas/community/core/service/tenant_manage/__init__.py
@@ -2,10 +2,8 @@
 
 from ._tenant_manage_service import (
     DefaultTenantManageService,
-    _record_to_response,
 )
 
 __all__ = [
     "DefaultTenantManageService",
-    "_record_to_response",
 ]
